microservices/user_details/requirements.txt
Flask==2.0.3
pymongo==3.12.1
psycopg2-binary==2.9.3 